move streamResponse out of RemoteLangGraphEventSource
