+++ CopilotKit/packages/runtime/src/agents/langgraph/event-source.ts
@@ -15,63 +15,7 @@ interface LangGraphEventWithState {
 }
 
 export class RemoteLangGraphEventSource {
-  private eventStream$ = new ReplaySubject<LangGraphEvent>();
-
-  async streamResponse(response: Response) {
-    const reader = response.body!.getReader();
-    const decoder = new TextDecoder();
-    let buffer = [];
-    const eventStream$ = this.eventStream$;
-
-    function flushBuffer() {
-      const currentBuffer = buffer.join("");
-      if (currentBuffer.trim().length === 0) {
-        return;
-      }
-      const parts = currentBuffer.split("\n");
-      if (parts.length === 0) {
-        return;
-      }
-
-      const lastPartIsComplete = currentBuffer.endsWith("\n");
-
-      // truncate buffer
-      buffer = [];
-
-      if (!lastPartIsComplete) {
-        // put back the last part
-        buffer.push(parts.pop());
-      }
-
-      parts
-        .map((part) => part.trim())
-        .filter((part) => part != "")
-        .forEach((part) => {
-          eventStream$.next(JSON.parse(part));
-        });
-    }
-
-    try {
-      while (true) {
-        const { done, value } = await reader.read();
-
-        if (!done) {
-          buffer.push(decoder.decode(value, { stream: true }));
-        }
-
-        flushBuffer();
-
-        if (done) {
-          break;
-        }
-      }
-    } catch (error) {
-      console.error("Error in stream", error);
-      eventStream$.error(error);
-      return;
-    }
-    eventStream$.complete();
-  }
+  public eventStream$ = new ReplaySubject<LangGraphEvent>();
 
   private shouldEmitToolCall(
     shouldEmitToolCalls: string | string[] | boolean,
