+++ CopilotKit/packages/runtime/src/lib/runtime/remote-actions.ts
@@ -5,9 +5,10 @@ import telemetry from "../../lib/telemetry-client";
 import { Message } from "../../graphql/types/converted";
 import { RuntimeEvent, RuntimeEventSubject } from "../../service-adapters/events";
 import { RemoteLangGraphEventSource } from "../../agents/langgraph/event-source";
-import { Observable } from "rxjs";
+import { Observable, ReplaySubject } from "rxjs";
 import { ActionInput } from "../../graphql/inputs/action.input";
 import { AgentStateInput } from "../../graphql/inputs/agent-state.input";
+import { LangGraphEvent } from "../../agents/langgraph/events";
 
 export type RemoteActionDefinition = {
   url: string;
@@ -213,12 +214,67 @@ function constructRemoteActions({
       }
 
       const eventSource = new RemoteLangGraphEventSource();
-      eventSource.streamResponse(response);
+      streamResponse(response, eventSource.eventStream$);
       return eventSource.processLangGraphEvents();
     },
   }));
 
   return [...actions, ...agents];
+}
+
+async function streamResponse(response: Response, eventStream$: ReplaySubject<LangGraphEvent>) {
+  const reader = response.body!.getReader();
+  const decoder = new TextDecoder();
+  let buffer = [];
+
+  function flushBuffer() {
+    const currentBuffer = buffer.join("");
+    if (currentBuffer.trim().length === 0) {
+      return;
+    }
+    const parts = currentBuffer.split("\n");
+    if (parts.length === 0) {
+      return;
+    }
+
+    const lastPartIsComplete = currentBuffer.endsWith("\n");
+
+    // truncate buffer
+    buffer = [];
+
+    if (!lastPartIsComplete) {
+      // put back the last part
+      buffer.push(parts.pop());
+    }
+
+    parts
+      .map((part) => part.trim())
+      .filter((part) => part != "")
+      .forEach((part) => {
+        eventStream$.next(JSON.parse(part));
+      });
+  }
+
+  try {
+    while (true) {
+      const { done, value } = await reader.read();
+
+      if (!done) {
+        buffer.push(decoder.decode(value, { stream: true }));
+      }
+
+      flushBuffer();
+
+      if (done) {
+        break;
+      }
+    }
+  } catch (error) {
+    console.error("Error in stream", error);
+    eventStream$.error(error);
+    return;
+  }
+  eventStream$.complete();
 }
 
 export async function setupRemoteActions({
